Swallow errors in RelayBroadcast
Previously, if forwarding of a broadcast frame resulted in a
FrameTooBigError, we would generate an ICMP "fragmentation needed"
packet by swapping the source and destination addresses.  But
this produced an ethernet frame with a broadcast source MAC, which
is problematic.

The Linux kernel never generates ICMP packets in response to
broadcasts, and it seems reasonable for us to do the same.

Futhermore, if we fail to relay a broadcast on one connection, that's
not a good reason to give up entirely - we can proceed to attempt to
relay on the other next hop connections.  Errors other than
FrameTooBigError that propagated upwards were simply logged, so that
logging is moved down into RelayBroadcast.